# cosmo: remove calls to reflect in FlatLCDM benchmarks