This CL removes all the calls to reflect in the FlatLCDM benchmarks.

```
$> benchstat ref-flatlcdm.txt new.txt
name                                  old time/op  new time/op  delta
FlatLCDMEN-4                          60.6µs ± 1%  60.6µs ± 2%     ~     (p=0.825 n=20+20)
FlatLCDMENdistance-4                  7.98ms ± 7%  0.17ms ± 1%  -97.88%  (p=0.000 n=20+20)
FlatLCDME-4                           6.42ns ± 2%  6.41ns ± 1%     ~     (p=0.572 n=20+20)
FlatLCDMEinv-4                        11.8ns ± 1%  11.7ns ± 0%   -0.47%  (p=0.003 n=20+16)
FlatLCDMComovingDistance-4            1.45µs ± 8%  0.60µs ± 1%  -58.83%  (p=0.000 n=20+20)
FlatLCDMComovingTransverseDistance-4  1.46µs ± 8%  0.61µs ± 1%  -58.52%  (p=0.000 n=20+18)
FlatLCDMLuminosityDistance-4          1.43µs ± 5%  0.62µs ± 2%  -56.79%  (p=0.000 n=20+20)
FlatLCDMLookbackTime-4                 240µs ± 1%   212µs ± 1%  -11.58%  (p=0.000 n=20+20)
FlatLCDMNComovingDistance-4           14.2ms ± 6%   6.2ms ± 1%  -56.57%  (p=0.000 n=19+20)
FlatLCDMNLuminosityDistance-4         14.5ms ± 7%   6.4ms ± 1%  -55.95%  (p=0.000 n=19+20)
FlatLCDMNE-4                          8.22ms ± 7%  0.17ms ± 1%  -97.94%  (p=0.000 n=20+20)
```